Refactor app bar
