+++ web/src/router.js
@@ -2,6 +2,8 @@ import Vue from 'vue';
 import Router from 'vue-router';
 
 import HomeView from '@/views/HomeView/HomeView.vue';
+import UserLoginView from '@/views/UserLoginView/UserLoginView.vue';
+import UserRegisterView from '@/views/UserRegisterView/UserRegisterView.vue';
 import PublicDandisetsView from '@/views/PublicDandisetsView/PublicDandisetsView.vue';
 import MyDandisetsView from '@/views/MyDandisetsView/MyDandisetsView.vue';
 import DandisetLandingView from '@/views/DandisetLandingView/DandisetLandingView.vue';
@@ -14,6 +16,16 @@ export default new Router({
       path: '/',
       name: 'home',
       component: HomeView,
+    },
+    {
+      path: '/user/login',
+      name: 'userLogin',
+      component: UserLoginView,
+    },
+    {
+      path: '/user/register',
+      name: 'userRegister',
+      component: UserRegisterView,
     },
     {
       path: '/dandiset',
