+++ web/src/views/UserLoginView/UserLoginView.vue
@@ -1,0 +1,21 @@
+<template>
+  <v-container>
+    TODO: UserLoginView
+    <GirderAuth
+      :force-otp="false"
+      :show-forgot-password="false"
+      :oauth="true"
+    />
+  </v-container>
+</template>
+
+<script>
+import { Authentication as GirderAuth } from '@girder/components/src/components';
+
+export default {
+  name: 'UserLoginView',
+  components: {
+    GirderAuth,
+  },
+};
+</script>
