+++ web/src/views/UserRegisterView/UserRegisterView.vue
@@ -1,0 +1,11 @@
+<template>
+  <v-container>
+    TODO: UserRegisterView
+  </v-container>
+</template>
+
+<script>
+export default {
+  name: 'UserRegisterView',
+};
+</script>
